Move common funct to general graphic utils
