+++ src/utils/graphic-layer-utils.js
@@ -1,3 +1,5 @@
+import { loadModules } from 'esri-loader';
+
 export const createGraphicLayer = (GraphicsLayer, graphic, id = "graphicLayer") => {
   return new GraphicsLayer({
     id: id,
@@ -16,6 +18,33 @@ export const createPointGraphic = (GraphicConstructor, symbol, lon, lat) => {
     geometry: point,
     symbol: symbol
   });
+}
+
+export const createGraphic = (Graphic, graphicStyles) => {
+  return new Graphic({
+    symbol: {
+      type: "polygon-3d",
+      symbolLayers: [
+        {
+          type: "fill",
+          material: { color: [...graphicStyles.colorRGB, graphicStyles.fillOpacity], },
+          outline: {
+            color: [...graphicStyles.colorRGB, graphicStyles.outlineOpacity],
+            size: graphicStyles.outlineWidth
+          }
+        }
+      ]
+    }
+  });
+}
+
+export const createPolygonGeometry = (gridCell) => {
+  return loadModules(["esri/geometry/Polygon"])
+    .then(([Polygon]) => {
+      return new Polygon(gridCell);
+    }).catch(error => {
+      console.error(error);
+    })
 }
 
 export const simplePictureMarker = asset => ({
