+++ src/utils/grid-layer-utils.js
@@ -1,25 +1,5 @@
 import { loadModules } from 'esri-loader';
 import { isEqual, flattenDeep } from 'lodash';
-import { gridCellDefaultStyles } from 'constants/landscape-view-constants';
-
-const { fillOpacity, outlineOpacity, outlineWidth, colorRGB } = gridCellDefaultStyles;
-export const createGraphic = (Graphic) => {
-  return new Graphic({
-    symbol: {
-      type: "polygon-3d",
-      symbolLayers: [
-        {
-          type: "fill",
-          material: { color: [...colorRGB, fillOpacity], },
-          outline: {
-            color: [...colorRGB, outlineOpacity],
-            size: outlineWidth
-          }
-        }
-      ]
-    }
-  });
-}
 
 export const calculateAgregatedGridCellGeometry = (hasContainedGridCells, gridCells, geometryEngineModule) => {
   if (!gridCells.length) return null;
@@ -40,15 +20,6 @@ export const calculateAggregatedCells = (features) => {
   return loadModules(["esri/geometry/geometryEngine"])
     .then(([geometryEngine]) => {
       return geometryEngine.union(features.map(gc => gc.geometry));
-    }).catch(error => {
-      console.error(error);
-    })
-}
-
-export const createPolygonGeometry = (gridCell) => {
-  return loadModules(["esri/geometry/Polygon"])
-    .then(([Polygon]) => {
-      return new Polygon(gridCell);
     }).catch(error => {
       console.error(error);
     })
